requirements.txt
absl-py==0.7.1
astor==0.8.0
atari-py==0.1.15
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
enum34==1.1.6
future==0.17.1
gast==0.2.2
gin-config==0.1.4
google-pasta==0.1.7
grpcio==1.21.1
gym==0.12.4
h5py==2.9.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
mock==3.0.5
numpy==1.16.3
opencv-python==4.1.0.25
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pyglet==1.3.2
Pygments==2.4.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1.1
scipy==1.3.0
six==1.12.0
tb-nightly==1.14.0a20190603
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==2.0.0b0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.1
yolk3k==0.9
